pol_env/Tribes/py/requirements.txt
py4j>=0.10.9.7
pillow>=10.0.0
